+++ module/actor/entity.js
@@ -9,20 +9,46 @@ export class WwnActor extends Actor {
   prepareData() {
     super.prepareData();
 
-    // Compute modifiers from actor scores
-    this.computeModifiers();
-    this.computeAC();
-    this.computeEncumbrance();
-    this._calculateMovement();
-    this.computeResources();
-    this.computeTreasure();
-    this.enableSpellcasting();
-    this.computeEffort();
-    this.computeSaves();
-    this.computeTotalSP();
-    this.setXP();
-    this.computePrepared();
-    this.computeInit();
+    if (this.type === "faction") {
+      const data = this.system;
+      const assets = (
+        this.items.filter((i) => i.type == "asset")
+      );
+      const cunningAssets = assets.filter(
+        (i) => i.system["assetType"] === "cunning"
+      );
+      const forceAssets = assets.filter(
+        (i) => i.system["assetType"] === "force"
+      );
+      const wealthAssets = assets.filter(
+        (i) => i.system["assetType"] === "wealth"
+      );
+  
+      data.cunningAssets = cunningAssets;
+      data.forceAssets = forceAssets;
+      data.wealthAssets = wealthAssets;
+  
+      data.health.max =
+        4 +
+        this.getHealth(data.wealthRating) +
+        this.getHealth(data.forceRating) +
+        this.getHealth(data.cunningRating);
+    } else {
+      // Compute modifiers from actor scores
+      this.computeModifiers();
+      this.computeAC();
+      this.computeEncumbrance();
+      this._calculateMovement();
+      this.computeResources();
+      this.computeTreasure();
+      this.enableSpellcasting();
+      this.computeEffort();
+      this.computeSaves();
+      this.computeTotalSP();
+      this.setXP();
+      this.computePrepared();
+      this.computeInit();
+    }
   }
 
   async createEmbeddedDocuments(embeddedName, data = [], context = {}) {
@@ -33,6 +59,15 @@ export class WwnActor extends Actor {
       }
     });
     super.createEmbeddedDocuments(embeddedName, data, context);
+  }
+
+  async _onCreate() {
+    if (this.type === "faction") {
+      await this.update({
+        "token.actorLink": true,
+        "img" : "systems/wwn/assets/default/faction.png"
+      });
+    }
   }
 
   /* -------------------------------------------- */
@@ -1129,7 +1164,31 @@ export class WwnActor extends Actor {
       await this.createEmbeddedDocuments("Item", skills);
     }
   }
+
+  // ----------------------------
+  // FACTION METHODS
+  // ----------------------------
+  getHealth(level) {
+    if (level in HEALTH__XP_TABLE) {
+      return HEALTH__XP_TABLE[level];
+    } else {
+      return 0;
+    }
+  }
+
 }
+
+export const HEALTH__XP_TABLE = {
+  1: 1,
+  2: 2,
+  3: 4,
+  4: 6,
+  5: 9,
+  6: 12,
+  7: 16,
+  8: 20,
+};
+
 
 function toCamelCase(text) {
   const split = text.split(" ").map((t) => t.titleCase());